Mon, Sep 11, 2023, 12:38 PM -07:00

+++ CS112/Assignment01/Part1/change_calc.h
@@ -10,6 +10,6 @@
 
 using namespace std;
 
-int change_calc(int, int, int, int);
+double change_calc(int, int, int, int);
 
 #endif
